+++ inst/modules/retractionwatch.R
@@ -56,7 +56,7 @@ retractionwatch <- function(paper) {
   # return a list ----
   list(
     table = table,
-    summary = summary_table,
+    summary_table = summary_table,
     na_replace = 0,
     traffic_light = tl,
     report = report,