+++ tests/testthat/test-08-module-power.R
@@ -13,17 +13,17 @@ test_that("power, no LLM", {
   expect_equal(mod_output$traffic_light, "na")
   expect_equal(nrow(mod_output$table), 0)
   expect_equal(nrow(mod_output$summary), 1)
-  expect_equal(mod_output$summary$power.n, 0)
-  expect_equal(mod_output$summary$power.complete, NA_integer_)
+  expect_equal(mod_output$summary_table$power.n, 0)
+  expect_equal(mod_output$summary_table$power.complete, NA_integer_)
 
   # several power sentences in one paragraph
   paper <- psychsci[[10]]
   mod_output <- module_run(paper, module)
   expect_equal(mod_output$traffic_light, "yellow")
   expect_equal(nrow(mod_output$table), 2)
-  expect_equal(nrow(mod_output$summary), 1)
-  expect_equal(mod_output$summary$power.n, 2)
-  expect_equal(mod_output$summary$power.complete, NA_integer_)
+  expect_equal(nrow(mod_output$summary_table), 1)
+  expect_equal(mod_output$summary_table$power.n, 2)
+  expect_equal(mod_output$summary_table$power.complete, NA_integer_)
 
   # multiple papers
   paper <- psychsci[10:15]
@@ -31,8 +31,8 @@ test_that("power, no LLM", {
   expect_equal(mod_output$traffic_light, "yellow")
   expect_equal(nrow(mod_output$table), 4)
   expect_equal(nrow(mod_output$summary), 6)
-  expect_equal(mod_output$summary$power.n, c(2, 0, 0, 0, 1, 1))
-  expect_equal(mod_output$summary$power.complete, rep(NA_integer_, 6))
+  expect_equal(mod_output$summary_table$power.n, c(2, 0, 0, 0, 1, 1))
+  expect_equal(mod_output$summary_table$power.complete, rep(NA_integer_, 6))
 })
 
 test_that("power, with LLM", {
@@ -48,8 +48,8 @@ test_that("power, with LLM", {
   expect_equal(mod_output$traffic_light, "na")
   expect_equal(nrow(mod_output$table), 0)
   expect_equal(nrow(mod_output$summary), 1)
-  expect_equal(mod_output$summary$power.n, 0)
-  expect_equal(mod_output$summary$power.complete, NA_integer_)
+  expect_equal(mod_output$summary_table$power.n, 0)
+  expect_equal(mod_output$summary_table$power.complete, NA_integer_)
 
   #httptest::start_capturing()
 
@@ -64,8 +64,8 @@ test_that("power, with LLM", {
     expect_equal(nrow(mod_output$table), 3)
     expect_equal(mod_output$table$sample, c(13500, 24, 72) |> as.character())
     expect_equal(nrow(mod_output$summary), 1)
-    expect_equal(mod_output$summary$power.n, 3)
-    expect_equal(mod_output$summary$power.complete, 2)
+    expect_equal(mod_output$summary_table$power.n, 3)
+    expect_equal(mod_output$summary_table$power.complete, 2)
 
     # multiple papers
     paper <- psychsci[10:15]
@@ -74,8 +74,8 @@ test_that("power, with LLM", {
     expect_equal(nrow(mod_output$table), 5)
     expect_equal(nrow(mod_output$summary), 6)
     expect_equal(mod_output$table$complete, c(F, T, T, F, F))
-    expect_equal(mod_output$summary$power.n, c(3, 0, 0, 0, 1, 1))
-    expect_equal(mod_output$summary$power.complete, c(2, NA, NA, NA, 0, 0))
+    expect_equal(mod_output$summary_table$power.n, c(3, 0, 0, 0, 1, 1))
+    expect_equal(mod_output$summary_table$power.complete, c(2, NA, NA, NA, 0, 0))
   })
 
   #httptest::stop_capturing()