+++ NEWS.md
@@ -1,6 +1,7 @@
 # metacheck 0.0.0.9060
 
 * Lots of changes for how reports are formatted
+* in module output, `summary` is now `summary_table`
 
 # metacheck 0.0.0.9059